+++ pitronically/users/forms.py
@@ -2,12 +2,21 @@ from django.contrib.auth import get_user_model, forms
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 
+from ckeditor_uploader.widgets import CKEditorUploadingWidget
+
 User = get_user_model()
 
 
 class UserChangeForm(forms.UserChangeForm):
+
     class Meta(forms.UserChangeForm.Meta):
         model = User
+        widgets = {
+            # 21.07.2019
+            # The introduction, which the user can enter will be a visual editor widget, so that it is more customizable
+            # especially in regards to whitespaces and line breaks as well as bold and italic editing
+            'introduction':     CKEditorUploadingWidget
+        }
 
 
 class UserCreationForm(forms.UserCreationForm):